export useDevice
